requirements.txt
streamlit
google-api-python-client
youtube-transcript-api
google-generativeai